+++ Fiaxe/Models/UploadItem.swift
@@ -8,14 +8,16 @@ struct UploadItem: Identifiable, Codable, Sendable {
     let r2Key: String
     let uploadDate: Date
     let publicURL: URL
+    let bucketName: String
 
-    init(fileName: String, fileSize: Int64, r2Key: String, publicURL: URL) {
+    init(fileName: String, fileSize: Int64, r2Key: String, publicURL: URL, bucketName: String) {
         self.id = UUID()
         self.fileName = fileName
         self.fileSize = fileSize
         self.r2Key = r2Key
         self.uploadDate = Date()
         self.publicURL = publicURL
+        self.bucketName = bucketName
     }
 
     var formattedFileSize: String {